requirements.txt
# python==3.10.6

pandas==2.2.2
openpyxl==3.1.5
torch==2.4.0
glob
tqdm
transformers==4.43.3
evaluate==0.4.2
nltk==3.9.1
rouge_score==0.1.2
bert-score==0.3.13

# for phi 3
tiktoken==0.7.0
einops==0.8.0
pytest==8.3.2
wheel==0.44.0
flash-attn==2.6.3

# for Orca 2
sentencepiece==0.2.0
protobuf==5.27.4

# for bluert
git+https://github.com/google-research/bleurt.git   # scipy==1.14.1 and tensorflow==2.17.0

# InfoLM
torchmetrics==1.4.1

# Close Models
google_generativeai==0.7.2
anthropic==0.34.1
openai==1.43.0